+++ src/main/java/net/mcreator/breadcraft/procedures/WildGreensGrassDropsProcedure.java
@@ -91,7 +91,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.061 && randomNumber <= 0.083) {
-								for (int index0 = 0; index0 < (int) (2); index0++) {
+								for (int index0 = 0; index0 < 2; index0++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -99,7 +99,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.103 && randomNumber <= 0.125) {
-								for (int index1 = 0; index1 < (int) (3); index1++) {
+								for (int index1 = 0; index1 < 3; index1++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -115,7 +115,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.025 && randomNumber <= 0.05) {
-								for (int index2 = 0; index2 < (int) (2); index2++) {
+								for (int index2 = 0; index2 < 2; index2++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -123,7 +123,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.05 && randomNumber <= 0.075) {
-								for (int index3 = 0; index3 < (int) (3); index3++) {
+								for (int index3 = 0; index3 < 3; index3++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -131,7 +131,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.075 && randomNumber <= 0.1) {
-								for (int index4 = 0; index4 < (int) (4); index4++) {
+								for (int index4 = 0; index4 < 4; index4++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -139,7 +139,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.075 && randomNumber <= 0.125) {
-								for (int index5 = 0; index5 < (int) (5); index5++) {
+								for (int index5 = 0; index5 < 5; index5++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -155,7 +155,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.017 && randomNumber <= 0.035) {
-								for (int index6 = 0; index6 < (int) (2); index6++) {
+								for (int index6 = 0; index6 < 2; index6++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -163,7 +163,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.035 && randomNumber <= 0.053) {
-								for (int index7 = 0; index7 < (int) (3); index7++) {
+								for (int index7 = 0; index7 < 3; index7++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -171,7 +171,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.053 && randomNumber <= 0.071) {
-								for (int index8 = 0; index8 < (int) (4); index8++) {
+								for (int index8 = 0; index8 < 4; index8++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -179,7 +179,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.071 && randomNumber <= 0.089) {
-								for (int index9 = 0; index9 < (int) (5); index9++) {
+								for (int index9 = 0; index9 < 5; index9++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -187,7 +187,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.089 && randomNumber <= 0.107) {
-								for (int index10 = 0; index10 < (int) (6); index10++) {
+								for (int index10 = 0; index10 < 6; index10++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -195,7 +195,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.107 && randomNumber <= 0.125) {
-								for (int index11 = 0; index11 < (int) (7); index11++) {
+								for (int index11 = 0; index11 < 7; index11++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -252,7 +252,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.061 && randomNumber <= 0.083) {
-								for (int index12 = 0; index12 < (int) (2); index12++) {
+								for (int index12 = 0; index12 < 2; index12++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -260,7 +260,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.103 && randomNumber <= 0.125) {
-								for (int index13 = 0; index13 < (int) (3); index13++) {
+								for (int index13 = 0; index13 < 3; index13++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -276,7 +276,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.025 && randomNumber <= 0.05) {
-								for (int index14 = 0; index14 < (int) (2); index14++) {
+								for (int index14 = 0; index14 < 2; index14++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -284,7 +284,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.05 && randomNumber <= 0.075) {
-								for (int index15 = 0; index15 < (int) (3); index15++) {
+								for (int index15 = 0; index15 < 3; index15++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -292,7 +292,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.075 && randomNumber <= 0.1) {
-								for (int index16 = 0; index16 < (int) (4); index16++) {
+								for (int index16 = 0; index16 < 4; index16++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -300,7 +300,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.075 && randomNumber <= 0.125) {
-								for (int index17 = 0; index17 < (int) (5); index17++) {
+								for (int index17 = 0; index17 < 5; index17++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -316,7 +316,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.017 && randomNumber <= 0.035) {
-								for (int index18 = 0; index18 < (int) (2); index18++) {
+								for (int index18 = 0; index18 < 2; index18++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -324,7 +324,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.035 && randomNumber <= 0.053) {
-								for (int index19 = 0; index19 < (int) (3); index19++) {
+								for (int index19 = 0; index19 < 3; index19++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -332,7 +332,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.053 && randomNumber <= 0.071) {
-								for (int index20 = 0; index20 < (int) (4); index20++) {
+								for (int index20 = 0; index20 < 4; index20++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -340,7 +340,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.071 && randomNumber <= 0.089) {
-								for (int index21 = 0; index21 < (int) (5); index21++) {
+								for (int index21 = 0; index21 < 5; index21++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -348,7 +348,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.089 && randomNumber <= 0.107) {
-								for (int index22 = 0; index22 < (int) (6); index22++) {
+								for (int index22 = 0; index22 < 6; index22++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -356,7 +356,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.107 && randomNumber <= 0.125) {
-								for (int index23 = 0; index23 < (int) (7); index23++) {
+								for (int index23 = 0; index23 < 7; index23++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -413,7 +413,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.041 && randomNumber <= 0.083) {
-								for (int index24 = 0; index24 < (int) (2); index24++) {
+								for (int index24 = 0; index24 < 2; index24++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -421,7 +421,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.083 && randomNumber <= 0.125) {
-								for (int index25 = 0; index25 < (int) (3); index25++) {
+								for (int index25 = 0; index25 < 3; index25++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -437,7 +437,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.025 && randomNumber <= 0.05) {
-								for (int index26 = 0; index26 < (int) (2); index26++) {
+								for (int index26 = 0; index26 < 2; index26++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -445,7 +445,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.05 && randomNumber <= 0.075) {
-								for (int index27 = 0; index27 < (int) (3); index27++) {
+								for (int index27 = 0; index27 < 3; index27++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -453,7 +453,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.075 && randomNumber <= 0.1) {
-								for (int index28 = 0; index28 < (int) (4); index28++) {
+								for (int index28 = 0; index28 < 4; index28++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -461,7 +461,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.075 && randomNumber <= 0.125) {
-								for (int index29 = 0; index29 < (int) (5); index29++) {
+								for (int index29 = 0; index29 < 5; index29++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -477,7 +477,7 @@ public class WildGreensGrassDropsProcedure {
 									_level.addFreshEntity(entityToSpawn);
 								}
 							} else if (randomNumber > 0.017 && randomNumber <= 0.035) {
-								for (int index30 = 0; index30 < (int) (2); index30++) {
+								for (int index30 = 0; index30 < 2; index30++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -485,7 +485,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.035 && randomNumber <= 0.053) {
-								for (int index31 = 0; index31 < (int) (3); index31++) {
+								for (int index31 = 0; index31 < 3; index31++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -493,7 +493,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.053 && randomNumber <= 0.071) {
-								for (int index32 = 0; index32 < (int) (4); index32++) {
+								for (int index32 = 0; index32 < 4; index32++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -501,7 +501,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.071 && randomNumber <= 0.089) {
-								for (int index33 = 0; index33 < (int) (5); index33++) {
+								for (int index33 = 0; index33 < 5; index33++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -509,7 +509,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.089 && randomNumber <= 0.107) {
-								for (int index34 = 0; index34 < (int) (6); index34++) {
+								for (int index34 = 0; index34 < 6; index34++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
@@ -517,7 +517,7 @@ public class WildGreensGrassDropsProcedure {
 									}
 								}
 							} else if (randomNumber > 0.107 && randomNumber <= 0.125) {
-								for (int index35 = 0; index35 < (int) (7); index35++) {
+								for (int index35 = 0; index35 < 7; index35++) {
 									if (world instanceof Level _level && !_level.isClientSide()) {
 										ItemEntity entityToSpawn = new ItemEntity(_level, (x + 0.5), (y + 0.5), (z + 0.5), itemSeeds);
 										entityToSpawn.setPickUpDelay(10);
